Add confirmation before closing multi-tab lxterminal
Close: #15
Add a confirmation dialog to ask user for sure before close a multi-tab
lxterminal. In order to add the confirmation, the signal handler of closing
window named terminal_close_window_activate_event need to be modified. I add
a dialog to ask users if they sure to close lxterminal or not. And I add a
new event handler named terminal_close_window_confirmation_event to handle
the delete-event signal, which will be emitted if a user requests that a
toplevel window is closed. There is a dialog same as mentioned above to
ask users if they sure to close this multi-tab lxterminal or not.
If user choose yes to the dialog, the handler will return FALSE and the
delete-event signal will be propagated further. On the contrary, if user
choose no to the dialog, the signal will be stop and no further action.