Add fields for XOFF and YOFF, including signs
Also make `geometry_columns` and `geometry_rows` unsigned integers, as `width` and `height` are unsigned integers in the `XParseGeometry` signature

Note that the signs of `XOFF` and `YOFF` are independent of the signs in the geometry specification.  The invocation `lxterminal --geometry=80x24+-100+-100` should position LXTerminal 100 pixels above and to the left of the top-left screen corner (i.e. off the screen).