fix: Fix VTE background echoes
This workaround is found in vte/src/app/app.cc, that avoiding letting
window painting the background can avoid background echoes.  However this
will cause other elements in the terminal window to have their background
transparent if not set.  To workaround this, we have to add background
color to the vbox of the window.