Add the "Uniform Lightness" colour palette
With the current default palette, VGA, the colours differ significantly
in lightness. Blue is especially dark and can therefore be difficult to
read on a black background.

This change adds a new palette where colours have a more uniform
lightness, and hand-picked hue and chroma values to make them
distinguishable from each other.
The foreground colour is white and not grey to increase the contrast
of most of the text shown in the terminal.