Add a new default colour palette
With the current default, VGA, the colours differ significantly in lightness.
Blue is especially dark and can therefore be difficult to read on a black background.

The new default colours have a more uniform lightness,
and hand-picked hue and chroma values to make them distinguishable from each other.
The foreground colour is white and not grey
to increase the contrast of most of the text shown in the terminal.